Refactor display logic into modules
